Extract welcome message in it's own file
