+++ .storybook/preview.js
@@ -1,37 +1,11 @@
 import '@atlaskit/css-reset';
 
 import React from 'react';
-import { colors } from '@atlaskit/theme';
 import { withPerformance } from 'storybook-addon-performance';
-import { version } from '../package.json';
 import GlobalStyles from './custom-decorators/global-styles';
+import welcomeMessage from "./welcome-message";
 
-// Doing this more complex check as console.table || console.log makes CI cry
-const consoleTable = Object.prototype.hasOwnProperty.call(console, 'table')
-  ? console.table
-  : console.log;
-
-// Generated by: http://patorjk.com/software/taag/#p=display&f=ANSI%20Shadow&t=rfd
-console.log(
-  `%c
-██████╗ ███████╗██████╗
-██╔══██╗██╔════╝██╔══██╗
-██████╔╝█████╗  ██║  ██║
-██╔══██╗██╔══╝  ██║  ██║
-██║  ██║██║     ██████╔╝
-╚═╝  ╚═╝╚═╝     ╚═════╝
-
-%cBeautiful and accessible drag and drop
-`,
-  `color: ${colors.G200}; font-size: 1.2em; font-weight: bold;`,
-  `color: ${colors.P200}; font-size: 1.2em; font-weight: bold;`,
-);
-
-consoleTable([
-  ['@react-forked/dnd version', version],
-  ['react version', React.version],
-  ['process.env.NODE_ENV', process.env.NODE_ENV],
-]);
+welcomeMessage();
 
 export const decorators = [
   (Story) => <GlobalStyles><Story/></GlobalStyles>,
