+++ .storybook/welcome-message.js
@@ -1,0 +1,34 @@
+import React from 'react';
+import { colors } from '@atlaskit/theme';
+import { version } from '../package.json';
+
+// Doing this more complex check as console.table || console.log makes CI cry
+const consoleTable = Object.prototype.hasOwnProperty.call(console, 'table')
+  ? console.table
+  : console.log;
+
+function welcomeMessage() {
+  // Generated by: http://patorjk.com/software/taag/#p=display&f=ANSI%20Shadow&t=rfd
+  console.log(
+    `%c
+  ██████╗ ███████╗██████╗
+  ██╔══██╗██╔════╝██╔══██╗
+  ██████╔╝█████╗  ██║  ██║
+  ██╔══██╗██╔══╝  ██║  ██║
+  ██║  ██║██║     ██████╔╝
+  ╚═╝  ╚═╝╚═╝     ╚═════╝
+
+  %cBeautiful and accessible drag and drop
+  `,
+    `color: ${colors.G200}; font-size: 1.2em; font-weight: bold;`,
+    `color: ${colors.P200}; font-size: 1.2em; font-weight: bold;`,
+  );
+
+  consoleTable([
+    ['@react-forked/dnd version', version],
+    ['react version', React.version],
+    ['process.env.NODE_ENV', process.env.NODE_ENV],
+  ]);
+}
+
+export default welcomeMessage;
